extract createFirebaseInstance
